+++ AUTHORS
@@ -1,10 +1,4 @@
 # This file lists all individuals having contributed content to the repository.
 # For how it is generated, see 'https://github.com/moul/rules.mk'
 
-allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
-dependabot-preview[bot] <27856297+dependabot-preview[bot]@users.noreply.github.com>
-dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
 Manfred Touron <94029+moul@users.noreply.github.com>
-Manfred Touron <m@42.am>
-moul-bot <bot@moul.io>
-Renovate Bot <bot@renovateapp.com>
