+++ README.md
@@ -5,11 +5,9 @@
 [![go.dev reference](https://img.shields.io/badge/go.dev-reference-007d9c?logo=go&logoColor=white)](https://pkg.go.dev/moul.io/progress)
 [![License](https://img.shields.io/badge/license-Apache--2.0%20%2F%20MIT-%2397ca00.svg)](https://github.com/moul/progress/blob/master/COPYRIGHT)
 [![GitHub release](https://img.shields.io/github/release/moul/progress.svg)](https://github.com/moul/progress/releases)
-[![Docker Metrics](https://images.microbadger.com/badges/image/moul/progress.svg)](https://microbadger.com/images/moul/progress)
 [![Made by Manfred Touron](https://img.shields.io/badge/made%20by-Manfred%20Touron-blue.svg?style=flat)](https://manfred.life/)
 
 [![Go](https://github.com/moul/progress/workflows/Go/badge.svg)](https://github.com/moul/progress/actions?query=workflow%3AGo)
-[![Release](https://github.com/moul/progress/workflows/Release/badge.svg)](https://github.com/moul/progress/actions?query=workflow%3ARelease)
 [![PR](https://github.com/moul/progress/workflows/PR/badge.svg)](https://github.com/moul/progress/actions?query=workflow%3APR)
 [![GolangCI](https://golangci.com/badges/github.com/moul/progress.svg)](https://golangci.com/r/github.com/moul/progress)
 [![codecov](https://codecov.io/gh/moul/progress/branch/master/graph/badge.svg)](https://codecov.io/gh/moul/progress)
@@ -20,16 +18,92 @@
 
 ## Usage
 
-[embedmd]:# (.tmp/usage.txt console)
-```console
-foo@bar:~$ progress hello world
-            _                                                   _                      _        _
- __ _  ___ | | __ _  _ _   __ _  ___  _ _  ___  _ __  ___  ___ | |_  ___  _ __   _ __ | | __ _ | |_  ___
-/ _` |/ _ \| |/ _` || ' \ / _` ||___|| '_|/ -_)| '_ \/ _ \|___||  _|/ -_)| '  \ | '_ \| |/ _` ||  _|/ -_)
-\__, |\___/|_|\__,_||_||_|\__, |     |_|  \___|| .__/\___/      \__|\___||_|_|_|| .__/|_|\__,_| \__|\___|
-|___/                     |___/                |_|                              |_|
-12 CPUs, /home/moul/go/bin/progress, fwrz, go1.15.5
-args ["progress","hello","world"]
+[embedmd]:# (example_test.go /import\ / $)
+```go
+import (
+	"fmt"
+
+	"moul.io/progress"
+	"moul.io/u"
+)
+
+func Example() {
+	// initialize a new progress.Progress
+	prog := progress.New()
+	prog.AddStep("init").SetDescription("initialize")
+	prog.AddStep("step1").SetDescription("step 1")
+	prog.AddStep("step2").SetData([]string{"hello", "world"}).SetDescription("step 2")
+	prog.AddStep("step3")
+	prog.AddStep("finish")
+
+	// automatically mark the last step as done when the function quit
+	defer prog.Get("finish").Done()
+
+	// mark init as Done
+	prog.Get("init").Done()
+
+	// mark step1 as started
+	prog.Get("step1").SetData(42).Start()
+
+	// then, mark it as done + attach custom data
+	prog.Get("step1").SetData(1337).Done()
+
+	// mark step2 as started
+	prog.Get("step2").Start()
+
+	fmt.Println(u.PrettyJSON(prog))
+
+	// outputs something like this:
+	// {
+	//  "steps": [
+	//    {
+	//      "id": "init",
+	//      "description": "initialize",
+	//      "started_at": "2020-12-22T20:26:05.717427484+01:00",
+	//      "done_at": "2020-12-22T20:26:05.717427484+01:00",
+	//      "state": "done"
+	//    },
+	//    {
+	//      "id": "step1",
+	//      "description": "step 1",
+	//      "started_at": "2020-12-22T20:26:05.71742797+01:00",
+	//      "done_at": "2020-12-22T20:26:05.717428258+01:00",
+	//      "state": "done",
+	//      "data": 1337,
+	//      "duration": 286
+	//    },
+	//    {
+	//      "id": "step2",
+	//      "description": "step 2",
+	//      "started_at": "2020-12-22T20:26:05.71742865+01:00",
+	//      "state": "in progress",
+	//      "data": [
+	//        "hello",
+	//        "world"
+	//      ],
+	//      "duration": 496251
+	//    },
+	//    {
+	//      "id": "step3"
+	//    },
+	//    {
+	//      "id": "finish"
+	//    }
+	//  ],
+	//  "created_at": "2020-12-22T20:26:05.717423018+01:00",
+	//  "snapshot": {
+	//    "state": "in progress",
+	//    "doing": "step 2",
+	//    "not_started": 2,
+	//    "in_progress": 1,
+	//    "completed": 2,
+	//    "total": 5,
+	//    "percent": 50,
+	//    "total_duration": 25935,
+	//    "started_at": "2020-12-22T20:26:05.717427484+01:00"
+	//  }
+	//}
+}
 ```
 
 ## Install
@@ -39,10 +113,6 @@ args ["progress","hello","world"]
 ```sh
 go get moul.io/progress
 ```
-
-### Releases
-
-See https://github.com/moul/progress/releases
 
 ## Contribute
