+++ go.mod
@@ -4,7 +4,6 @@ go 1.13
 
 require (
 	github.com/stretchr/testify v1.6.1
-	github.com/tailscale/depaware v0.0.0-20201210233412-71b54857b5d9
-	golang.org/x/tools v0.0.0-20201005185003-576e169c3de7 // indirect
+	github.com/tailscale/depaware v0.0.0-20201214215404-77d1e9757027
 	moul.io/u v1.19.1
 )
